# Company-Eod-Site/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: EOD Management Portal API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated

  # ── USERS ─────────────────────────────────────────────────────────────────
...
  /daily-work:
    get:
      operationId: listDailyWork
      tags: [dailyWork]
      summary: List daily work items
      parameters:
        - name: userId
          in: query
          schema:
            type: ["integer", "null"]
        - name: teamId
          in: query
          schema:
            type: ["integer", "null"]
        - name: date
          in: query
          schema:
            type: ["string", "null"]
        - name: month
          in: query
          schema:
            type: ["integer", "null"]
        - name: year
          in: query
          schema:
            type: ["integer", "null"]
        - name: status
          in: query
          schema:
            type: ["string", "null"]
        - name: week
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: Daily work items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DailyWorkItem"
    post:
      operationId: createDailyWork
      tags: [dailyWork]
      summary: Add daily work item
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DailyWorkInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DailyWorkItem"
  /daily-work/{id}:
    get:
      operationId: getDailyWork
      tags: [dailyWork]
      summary: Get daily work item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Daily work item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DailyWorkItem"
    patch:
      operationId: updateDailyWork
      tags: [dailyWork]
      summary: Update daily work item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DailyWorkUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DailyWorkItem"
    delete:
      operationId: deleteDailyWork
      tags: [dailyWork]
      summary: Delete daily work item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── TRAINING ──────────────────────────────────────────────────────────────
...
  /dashboard/stats:
    get:
      operationId: getDashboardStats
      tags: [dashboard]
      summary: Get dashboard statistics
      parameters:
        - name: date
          in: query
          schema:
            type: ["string", "null"]
        - name: teamId
          in: query
          schema:
            type: ["integer", "null"]
        - name: month
          in: query
          schema:
            type: ["integer", "null"]
        - name: year
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: Dashboard stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardStats"
  /dashboard/team-summary:
    get:
      operationId: getTeamSummary
      tags: [dashboard]
      summary: Team-wise EOD completion summary
      parameters:
        - name: date
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: Team summary
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TeamSummary"
  /dashboard/monthly-trend:
    get:
      operationId: getMonthlyTrend
      tags: [dashboard]
      summary: Monthly EOD and task completion trend
      parameters:
        - name: year
          in: query
          schema:
            type: ["integer", "null"]
        - name: teamId
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: Monthly trend
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MonthlyTrend"

  # ── NOTIFICATIONS ─────────────────────────────────────────────────────────
...
components:
  schemas:
...
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
    AuthResponse:
      type: object
      required: [user, token]
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string

    # ── USER ──────────────────────────────────────────────────────────────
    User:
      type: object
      required: [id, name, email, role]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          description: "employee | tl | manager | hr"
        teamId:
          type: ["integer", "null"]
        teamName:
          type: ["string", "null"]
        employeeId:
          type: ["string", "null"]
        department:
          type: ["string", "null"]
        createdAt:
          type: string
...
    DailyWorkItem:
      type: object
      required: [id, action, date, status]
      properties:
        id:
          type: integer
        action:
          type: string
        how:
          type: ["string", "null"]
        who:
          type: ["string", "null"]
        date:
          type: string
        startDate:
          type: ["string", "null"]
        completionDate:
          type: ["string", "null"]
        status:
          type: string
          description: "completed | wip | yts | hold"
        completionPct:
          type: ["integer", "null"]
        remarks:
          type: ["string", "null"]
        userId:
          type: ["integer", "null"]
        teamId:
          type: ["integer", "null"]
        userName:
          type: ["string", "null"]
        teamName:
          type: ["string", "null"]
        createdAt:
          type: string
    DailyWorkInput:
      type: object
      required: [action, date, status]
      properties:
        action:
          type: string
        how:
          type: ["string", "null"]
        who:
          type: ["string", "null"]
        date:
          type: string
        startDate:
          type: ["string", "null"]
        completionDate:
          type: ["string", "null"]
        status:
          type: string
        completionPct:
          type: ["integer", "null"]
        remarks:
          type: ["string", "null"]
        userId:
          type: ["integer", "null"]
        teamId:
          type: ["integer", "null"]
    DailyWorkUpdate:
      type: object
      properties:
        action:
          type: string
        how:
          type: ["string", "null"]
        who:
          type: ["string", "null"]
        date:
          type: string
        startDate:
          type: ["string", "null"]
        completionDate:
          type: ["string", "null"]
        status:
          type: string
        completionPct:
          type: ["integer", "null"]
        remarks:
          type: ["string", "null"]

    # ── TRAINING ──────────────────────────────────────────────────────────
...
    DashboardStats:
      type: object
      required: [totalEmployees, submittedEod, pendingEod, absentEmployees, onLeave, trainingAttended, completedTasks, wipTasks, holdTasks, ytsTasks, teamCount]
      properties:
        totalEmployees:
          type: integer
        submittedEod:
          type: integer
        pendingEod:
          type: integer
        absentEmployees:
          type: integer
        onLeave:
          type: integer
        trainingAttended:
          type: integer
        completedTasks:
          type: integer
        wipTasks:
          type: integer
        holdTasks:
          type: integer
        ytsTasks:
          type: integer
        teamCount:
          type: integer
        completionRate:
          type: ["number", "null"]
    TeamSummary:
      type: object
      required: [teamId, teamName, totalMembers, submitted, pending, absent]
      properties:
        teamId:
          type: integer
        teamName:
          type: string
        totalMembers:
          type: integer
        submitted:
          type: integer
        pending:
          type: integer
        absent:
          type: integer
        completionPct:
          type: number
    MonthlyTrend:
      type: object
      required: [month, year, submitted, pending, totalTasks, completedTasks]
      properties:
        month:
          type: integer
        year:
          type: integer
        monthLabel:
          type: string
        submitted:
          type: integer
        pending:
          type: integer
        totalTasks:
          type: integer
        completedTasks:
          type: integer

    # ── NOTIFICATIONS ─────────────────────────────────────────────────────
...