# Company-Eod-Site/lib/api-zod/src/generated/types/authResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * EOD Management Portal API
 * OpenAPI spec version: 0.1.0
 */
import type { User } from './user';

export interface AuthResponse {
  user: User;
  token: string;
}